Issue #1405: DataStorageControllerTest split into 3 test classes
