+++ api/src/test/java/com/epam/pipeline/controller/datastorage/DataStorageControllerTest.java
@@ -7,11 +7,11 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.epam.pipeline.controller.datastorage;
@@ -19,19 +19,12 @@ package com.epam.pipeline.controller.datastorage;
 import com.epam.pipeline.config.JsonMapper;
 import com.epam.pipeline.controller.Result;
 import com.epam.pipeline.controller.vo.DataStorageVO;
-import com.epam.pipeline.controller.vo.GenerateDownloadUrlVO;
 import com.epam.pipeline.controller.vo.UploadFileMetadata;
-import com.epam.pipeline.controller.vo.data.storage.UpdateDataStorageItemVO;
 import com.epam.pipeline.controller.vo.security.EntityWithPermissionVO;
 import com.epam.pipeline.entity.SecuredEntityWithAction;
 import com.epam.pipeline.entity.datastorage.AbstractDataStorage;
-import com.epam.pipeline.entity.datastorage.ContentDisposition;
 import com.epam.pipeline.entity.datastorage.DataStorageAction;
-import com.epam.pipeline.entity.datastorage.DataStorageDownloadFileUrl;
 import com.epam.pipeline.entity.datastorage.DataStorageFile;
-import com.epam.pipeline.entity.datastorage.DataStorageFolder;
-import com.epam.pipeline.entity.datastorage.DataStorageItemContent;
-import com.epam.pipeline.entity.datastorage.DataStorageListing;
 import com.epam.pipeline.entity.datastorage.DataStorageStreamingContent;
 import com.epam.pipeline.entity.datastorage.DataStorageWithShareMount;
 import com.epam.pipeline.entity.datastorage.PathDescription;
@@ -40,129 +33,30 @@ import com.epam.pipeline.entity.datastorage.StorageUsage;
 import com.epam.pipeline.entity.datastorage.TemporaryCredentials;
 import com.epam.pipeline.entity.datastorage.aws.S3bucketDataStorage;
 import com.epam.pipeline.entity.datastorage.rules.DataStorageRule;
-import com.epam.pipeline.manager.datastorage.DataStorageApiService;
 import com.epam.pipeline.test.creator.datastorage.DatastorageCreatorUtils;
-import com.epam.pipeline.test.web.AbstractControllerTest;
 import com.fasterxml.jackson.core.type.TypeReference;
-import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
 import java.io.InputStream;
-import java.nio.charset.Charset;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
-import static com.epam.pipeline.test.creator.CommonCreatorConstants.INTEGER_TYPE;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.OBJECT_TYPE;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.STRING_STRING_MAP_TYPE;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.STRING_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
-@WebMvcTest(controllers = DataStorageController.class)
-public class DataStorageControllerTest extends AbstractControllerTest {
-
-    private static final long ID = 1L;
-    private static final int TEST_INT = 1;
-    private static final String TEST = "TEST";
-    private static final String DATASTORAGE_URL = SERVLET_PATH + "/datastorage";
-    private static final String LOAD_ALL_URL = DATASTORAGE_URL + "/loadAll";
-    private static final String LOAD_AVAILABLE_URL = DATASTORAGE_URL + "/available";
-    private static final String LOAD_AVAILABLE_WITH_MOUNTS = DATASTORAGE_URL + "/availableWithMounts";
-    private static final String LOAD_WRITABLE_URL = DATASTORAGE_URL + "/mount";
-    private static final String BY_ID_URL = DATASTORAGE_URL + "/%d";
-    private static final String LOAD_DATASTORAGE = BY_ID_URL + "/load";
-    private static final String FIND_URL = DATASTORAGE_URL + "/find";
-    private static final String FIND_BY_PATH_URL = DATASTORAGE_URL + "/findByPath";
-    private static final String DATASTORAGE_ITEMS_URL = BY_ID_URL + "/list";
-    private static final String DATASTORAGE_LISTING_URL = DATASTORAGE_ITEMS_URL + "/page";
-    private static final String DATASTORAGE_ITEMS_UPLOAD_URL = DATASTORAGE_ITEMS_URL + "/upload";
-    private static final String DATASTORAGE_UPLOAD_STREAM_URL = BY_ID_URL + "/upload/stream";
-    private static final String DATASTORAGE_DOWNLOAD_STREAM_URL = BY_ID_URL + "/download";
-    private static final String DATASTORAGE_ITEMS_CONTENT = BY_ID_URL + "/content";
-    private static final String DOWNLOAD_REDIRECT_URL = BY_ID_URL + "/downloadRedirect";
-    private static final String GENERATE_URL = BY_ID_URL + "/generateUrl";
-    private static final String GENERATE_UPLOAD_URL = BY_ID_URL + "/generateUploadUrl";
-    private static final String CONTENT_URL = BY_ID_URL + "/content";
-    private static final String RESTORE_VERSION_URL = DATASTORAGE_ITEMS_URL + "/restore";
-    private static final String DATASTORAGE_SAVE_URL= DATASTORAGE_URL + "/save";
-    private static final String DATASTORAGE_UPDATE_URL = DATASTORAGE_URL + "/update";
-    private static final String DATASTORAGE_POLICY_URL = DATASTORAGE_URL + "/policy";
-    private static final String DATASTORAGE_DELETE_URL = BY_ID_URL + "/delete";
-    private static final String DATASTORAGE_RULE_URL = DATASTORAGE_URL + "/rule";
-    private static final String SAVE_RULE_URL = DATASTORAGE_RULE_URL + "/register";
-    private static final String LOAD_RULES_URL = DATASTORAGE_RULE_URL + "/load";
-    private static final String DELETE_RULES_URL = DATASTORAGE_RULE_URL + "/delete";
-    private static final String TEMP_CREDENTIALS_URL = DATASTORAGE_URL + "/tempCredentials/";
-    private static final String TAGS_URL = BY_ID_URL + "/tags";
-    private static final String TAGS_LIST_URL = TAGS_URL + "/list";
-    private static final String SHARED_LINK_URL = BY_ID_URL + "/sharedLink";
-    private static final String PERMISSION_URL = DATASTORAGE_URL + "/permission";
-    private static final String PATH_URL = DATASTORAGE_URL + "/path";
-    private static final String PATH_SIZE_URL = PATH_URL + "/size";
-    private static final String PATH_USAGE_URL = PATH_URL + "/usage";
-    private static final String SHARED_STORAGE_URL = DATASTORAGE_URL + "/sharedStorage";
-    private static final String OCTET_STREAM_CONTENT_TYPE = "application/octet-stream";
-    private static final String ID_AS_STRING = String.valueOf(ID);
-    private static final String TRUE_AS_STRING = String.valueOf(true);
-    private static final String FALSE_AS_STRING = String.valueOf(false);
-    private static final String FROM_REGION = "fromRegion";
-    private static final String ID_PARAM = "id";
-    private static final String PATH = "path";
-    private static final String SHOW_VERSION = "showVersion";
-    private static final String PAGE_SIZE = "pageSize";
-    private static final String MARKER = "marker";
-    private static final String VERSION = "version";
-    private static final String TOTALLY = "totally";
-    private static final String TEST_PATH = "localhost:root/";
-    private static final String CONTENT_DISPOSITION_PARAM = "contentDisposition";
-    private static final String CONTENT_DISPOSITION_HEADER = "Content-Disposition";
-    private static final String CLOUD = "cloud";
-    private static final String SKIP_POLICY = "skipPolicy";
-    private static final String PIPELINE_ID = "pipelineId";
-    private static final String FILE_MASK = "fileMask";
-    private static final String FILTER_MASK = "filterMask";
-    private static final String REWRITE = "rewrite";
-    private static final String EMPTY_STRING = "";
-    private static final String PAGE = "page";
-    private static final String RUN_ID = "runId";
-    private static final String FILE_NAME = "file.txt";
-    private static final String FILE_SIZE = "0 Kb";
-    private static final String MULTIPART_CONTENT_TYPE =
-            "multipart/form-data; boundary=--------------------------boundary";
-    private static final String MULTIPART_CONTENT = "----------------------------boundary\r\n" +
-            "Content-Disposition: form-data; name=\"file\"; filename=\"file.txt\"\r\n" +
-            "Content-Type: application/octet-stream\r\n" +
-            "\r\n" +
-            TEST +
-            "\r\n" +
-            "----------------------------boundary--";
-    private static final ContentDisposition CONTENT_DISPOSITION = ContentDisposition.INLINE;
-    private static final Map<String, String> TAGS = Collections.singletonMap(TEST, TEST);
-    private final DataStorageRule dataStorageRule = DatastorageCreatorUtils.getDataStorageRule();
-    private final S3bucketDataStorage s3Bucket = DatastorageCreatorUtils.getS3bucketDataStorage();
-    private final DataStorageFile file = DatastorageCreatorUtils.getDataStorageFile();
-    private final DataStorageFolder folder = DatastorageCreatorUtils.getDataStorageFolder();
-    private final UpdateDataStorageItemVO update = DatastorageCreatorUtils.getUpdateDataStorageItemVO();
-    private final List<Pair<AbstractDataStorage, TypeReference>> storageTypeReferenceList =
-            DatastorageCreatorUtils.getRegularTypeStorages();
-    private final List<Pair<AbstractDataStorage, TypeReference>> securedStorageTypeReferenceList =
-            DatastorageCreatorUtils.getSecuredTypeStorages();
-
-    @Autowired
-    private DataStorageApiService mockStorageApiService;
+public class DataStorageControllerTest extends AbstractDataStorageControllerTest {
 
     @Test
     public void shouldFaiGetDataStoragesForUnauthorizedUser() {
@@ -249,89 +143,6 @@ public class DataStorageControllerTest extends AbstractControllerTest {
     }
 
     @Test
-    public void shouldFailGetDataStorageItemsForUnauthorizedUser() {
-        performUnauthorizedRequest(get(String.format(DATASTORAGE_ITEMS_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageFolder() {
-        final DataStorageListing dataStorageListing = DatastorageCreatorUtils.getDataStorageListing();
-        final List<DataStorageFolder> folders = Collections.singletonList(folder);
-        dataStorageListing.setResults(Collections.singletonList(folder));
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST);
-        params.add(SHOW_VERSION, FALSE_AS_STRING);
-        Mockito.doReturn(dataStorageListing)
-                .when(mockStorageApiService).getDataStorageItems(ID, TEST, false, null, null);
-
-        final MvcResult mvcResult = performRequest(get(String.format(DATASTORAGE_ITEMS_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItems(ID, TEST, false, null, null);
-        assertResponse(mvcResult, folders, DatastorageCreatorUtils.DATA_STORAGE_FOLDER_LIST_TYPE);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageOwnerFile() {
-        final DataStorageListing dataStorageListing = DatastorageCreatorUtils.getDataStorageListing();
-        final List<DataStorageFile> files = Collections.singletonList(file);
-        dataStorageListing.setResults(Collections.singletonList(file));
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST);
-        params.add(SHOW_VERSION, TRUE_AS_STRING);
-        Mockito.doReturn(dataStorageListing)
-                .when(mockStorageApiService).getDataStorageItemsOwner(ID, TEST, true, null, null);
-
-        final MvcResult mvcResult = performRequest(get(String.format(DATASTORAGE_ITEMS_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemsOwner(ID, TEST, true, null, null);
-        assertResponse(mvcResult, files, DatastorageCreatorUtils.DATA_STORAGE_FILE_LIST_TYPE);
-    }
-
-    @Test
-    public void shouldFailGetDataStorageItemsListingForUnauthorizedUser() {
-        performUnauthorizedRequest(get(String.format(DATASTORAGE_LISTING_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageItems() {
-        final DataStorageListing dataStorageListing = DatastorageCreatorUtils.getDataStorageListing();
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST);
-        params.add(SHOW_VERSION, FALSE_AS_STRING);
-        params.add(PAGE_SIZE, ID_AS_STRING);
-        params.add(MARKER, TEST);
-        Mockito.doReturn(dataStorageListing)
-                .when(mockStorageApiService).getDataStorageItems(ID, TEST, false, TEST_INT, TEST);
-
-        final MvcResult mvcResult = performRequest(get(String.format(DATASTORAGE_LISTING_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItems(ID, TEST, false, TEST_INT, TEST);
-        assertResponse(mvcResult, dataStorageListing, DatastorageCreatorUtils.DATA_STORAGE_LISTING_TYPE);
-    }
-
-    @Test
-    public void shouldFailUpdateDataStorageItemsForUnauthorizedUser() {
-        performUnauthorizedRequest(post(String.format(DATASTORAGE_ITEMS_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldUpdateDataStorageItems() throws Exception {
-        final List<UpdateDataStorageItemVO> updateList = Collections.singletonList(update);
-        final List<DataStorageFile> files = Collections.singletonList(file);
-        final String content = getObjectMapper().writeValueAsString(updateList);
-        Mockito.doReturn(files).when(mockStorageApiService).updateDataStorageItems(ID, updateList);
-
-        final MvcResult mvcResult = performRequest(post(String.format(DATASTORAGE_ITEMS_URL, ID)).content(content));
-
-        Mockito.verify(mockStorageApiService).updateDataStorageItems(ID, updateList);
-        assertResponse(mvcResult, files, DatastorageCreatorUtils.DATA_STORAGE_FILE_LIST_TYPE);
-    }
-
-    @Test
     public void shouldFailUploadFileForUnauthorizedUser() {
         performUnauthorizedRequest(post(String.format(DATASTORAGE_ITEMS_UPLOAD_URL, ID)));
     }
@@ -353,7 +164,7 @@ public class DataStorageControllerTest extends AbstractControllerTest {
         Mockito.verify(mockStorageApiService).createDataStorageFile(ID, TEST, FILE_NAME, TEST.getBytes());
         final List<UploadFileMetadata> actualResult = JsonMapper
                 .parseData(mvcResult.getResponse().getContentAsString(),
-                           new TypeReference<List<UploadFileMetadata>>() { });
+                        new TypeReference<List<UploadFileMetadata>>() { });
         Assert.assertEquals(uploadFileMetadataList, actualResult);
     }
 
@@ -368,14 +179,14 @@ public class DataStorageControllerTest extends AbstractControllerTest {
         final List<DataStorageFile> dataStorageFiles = Collections.singletonList(file);
         Mockito.doReturn(file).when(mockStorageApiService)
                 .createDataStorageFile(Mockito.eq(ID), Mockito.eq(TEST),
-                                       Mockito.eq(FILE_NAME), (InputStream) Mockito.any());
+                        Mockito.eq(FILE_NAME), (InputStream) Mockito.any());
 
         final MvcResult mvcResult = performRequest(post(String.format(DATASTORAGE_UPLOAD_STREAM_URL, ID))
                 .content(MULTIPART_CONTENT).param(PATH, TEST), MULTIPART_CONTENT_TYPE, EXPECTED_CONTENT_TYPE);
 
         Mockito.verify(mockStorageApiService)
                 .createDataStorageFile(Mockito.eq(ID), Mockito.eq(TEST),
-                                       Mockito.eq(FILE_NAME), (InputStream) Mockito.any());
+                        Mockito.eq(FILE_NAME), (InputStream) Mockito.any());
         assertResponse(mvcResult, dataStorageFiles, new TypeReference<Result<List<DataStorageFile>>>() { });
     }
 
@@ -403,225 +214,6 @@ public class DataStorageControllerTest extends AbstractControllerTest {
         final String contentDispositionHeader = mvcResult.getResponse().getHeader(CONTENT_DISPOSITION_HEADER);
         Assert.assertEquals(TEST, actualData);
         assertThat(contentDispositionHeader).contains(TEST);
-    }
-
-    @Test
-    public void shouldFailUploadStorageItemForUnauthorizedUser() {
-        performUnauthorizedRequest(post(String.format(DATASTORAGE_ITEMS_CONTENT, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldUploadStorageItem() throws Exception {
-        Mockito.doReturn(file).when(mockStorageApiService)
-                .createDataStorageFile(ID, TEST, TEST.getBytes(Charset.defaultCharset()));
-        final String content = getObjectMapper().writeValueAsString(TEST);
-
-        final MvcResult mvcResult = performRequest(post(String.format(DATASTORAGE_ITEMS_CONTENT, ID))
-                .param(PATH, TEST).content(content));
-
-        Mockito.verify(mockStorageApiService).createDataStorageFile(ID, TEST, TEST.getBytes(Charset.defaultCharset()));
-        assertResponse(mvcResult, file, DatastorageCreatorUtils.DATA_STORAGE_FILE_TYPE);
-    }
-
-    @Test
-    public void shouldFailDeleteDataStorageItemForUnauthorizedUser() {
-        performUnauthorizedRequest(delete(String.format(DATASTORAGE_ITEMS_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDeleteDataStorageItems() throws Exception {
-        final List<UpdateDataStorageItemVO> updateList = Collections.singletonList(update);
-        final String content = getObjectMapper().writeValueAsString(updateList);
-        Mockito.doReturn(TEST_INT).when(mockStorageApiService).deleteDataStorageItems(ID, updateList, false);
-
-        final MvcResult mvcResult = performRequest(delete(String.format(DATASTORAGE_ITEMS_URL, ID))
-                .param(TOTALLY, FALSE_AS_STRING).content(content));
-
-        Mockito.verify(mockStorageApiService).deleteDataStorageItems(ID, updateList, false);
-        assertResponse(mvcResult, TEST_INT, INTEGER_TYPE);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDeleteDataStorageItemsOwner() throws Exception {
-        final List<UpdateDataStorageItemVO> updateList = Collections.singletonList(update);
-        final String content = getObjectMapper().writeValueAsString(updateList);
-        Mockito.doReturn(TEST_INT).when(mockStorageApiService).deleteDataStorageItemsOwner(ID, updateList, true);
-
-        final MvcResult mvcResult = performRequest(delete(String.format(DATASTORAGE_ITEMS_URL, ID))
-                .param(TOTALLY, TRUE_AS_STRING).content(content));
-
-        Mockito.verify(mockStorageApiService).deleteDataStorageItemsOwner(ID, updateList, true);
-        assertResponse(mvcResult, TEST_INT, INTEGER_TYPE);
-    }
-
-    @Test
-    public void shouldFailGenerateItemUrlAndRedirectForUnauthorizedUser() {
-        performUnauthorizedRequest(get(String.format(DOWNLOAD_REDIRECT_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateItemUrlAndRedirect() {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST_PATH);
-        params.add(CONTENT_DISPOSITION_PARAM, CONTENT_DISPOSITION.name());
-        Mockito.doReturn(url).when(mockStorageApiService)
-                .generateDataStorageItemUrl(ID, TEST_PATH, null, CONTENT_DISPOSITION);
-
-        final MvcResult mvcResult = performRedirectedRequest(
-                get(String.format(DOWNLOAD_REDIRECT_URL, ID)).params(params), TEST
-        );
-
-        Mockito.verify(mockStorageApiService).generateDataStorageItemUrl(ID, TEST_PATH, null, CONTENT_DISPOSITION);
-        final String actual = mvcResult.getResponse().getRedirectedUrl();
-        assertThat(actual).isEqualTo(TEST);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateItemUrlAndRedirectOwner() {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST_PATH);
-        params.add(VERSION, TEST);
-        params.add(CONTENT_DISPOSITION_PARAM, CONTENT_DISPOSITION.name());
-        Mockito.doReturn(url).when(mockStorageApiService)
-                .generateDataStorageItemUrlOwner(ID, TEST_PATH, TEST, CONTENT_DISPOSITION);
-
-        final MvcResult mvcResult = performRedirectedRequest(get(String.format(DOWNLOAD_REDIRECT_URL, ID))
-                .params(params), TEST);
-
-        Mockito.verify(mockStorageApiService).generateDataStorageItemUrlOwner(ID, TEST_PATH, TEST, CONTENT_DISPOSITION);
-        final String actual = mvcResult.getResponse().getRedirectedUrl();
-        assertThat(actual).isEqualTo(TEST);
-    }
-
-    @Test
-    public void shouldFailGenerateDataStorageItemUrl() {
-        performUnauthorizedRequest(get(String.format(GENERATE_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateDataStorageItemUrlOwner() {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST_PATH);
-        params.add(VERSION, TEST);
-        params.add(CONTENT_DISPOSITION_PARAM, CONTENT_DISPOSITION.name());
-        Mockito.doReturn(url).when(mockStorageApiService)
-                .generateDataStorageItemUrlOwner(ID, TEST_PATH, TEST, CONTENT_DISPOSITION);
-
-        final MvcResult mvcResult = performRequest(get(String.format(GENERATE_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService)
-                .generateDataStorageItemUrlOwner(ID, TEST_PATH, TEST, CONTENT_DISPOSITION);
-
-        assertResponse(mvcResult, url, DatastorageCreatorUtils.DOWNLOAD_FILE_URL_TYPE);
-    }
-
-    @Test
-    public void shouldFailGenerateDataStorageItemUploadUrlForUnauthorizedUser() {
-        performUnauthorizedRequest(get(String.format(GENERATE_UPLOAD_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateDataStorageItemUploadUrl() {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        Mockito.doReturn(url).when(mockStorageApiService).generateDataStorageItemUploadUrl(ID, TEST);
-
-        final MvcResult mvcResult = performRequest(get(String.format(GENERATE_UPLOAD_URL, ID)).param(PATH, TEST));
-
-        Mockito.verify(mockStorageApiService).generateDataStorageItemUploadUrl(ID, TEST);
-        assertResponse(mvcResult, url, DatastorageCreatorUtils.DOWNLOAD_FILE_URL_TYPE);
-    }
-
-    @Test
-    public void shouldFailGetDataStorageItemContentForUnauthorizedUser() {
-        performUnauthorizedRequest(get(String.format(CONTENT_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageItemContent() {
-        final DataStorageItemContent dataStorageItemContent = DatastorageCreatorUtils.getDataStorageItemContent();
-        Mockito.doReturn(dataStorageItemContent).when(mockStorageApiService)
-                .getDataStorageItemContent(ID, TEST, null);
-
-        final MvcResult mvcResult = performRequest(get(String.format(CONTENT_URL, ID)).param(PATH, TEST));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemContent(ID, TEST, null);
-        assertResponse(mvcResult, dataStorageItemContent, DatastorageCreatorUtils.ITEM_CONTENT_URL);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageItemContentOwner() {
-        final DataStorageItemContent dataStorageItemContent = DatastorageCreatorUtils.getDataStorageItemContent();
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(PATH, TEST);
-        params.add(VERSION, TEST);
-        Mockito.doReturn(dataStorageItemContent).when(mockStorageApiService)
-                .getDataStorageItemContentOwner(ID, TEST, TEST);
-
-        final MvcResult mvcResult = performRequest(get(String.format(CONTENT_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemContentOwner(ID, TEST, TEST);
-        assertResponse(mvcResult, dataStorageItemContent, DatastorageCreatorUtils.ITEM_CONTENT_URL);
-    }
-
-    @Test
-    public void shouldFailGenerateDataStorageItemsUrlsForUnauthorizedUser() {
-        performUnauthorizedRequest(post(String.format(GENERATE_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateDataStorageItemsUrls() throws Exception {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        final List<DataStorageDownloadFileUrl> urls = Collections.singletonList(url);
-        final GenerateDownloadUrlVO generateDownloadUrlVO = DatastorageCreatorUtils.getGenerateDownloadUrlVO();
-        final String content = getObjectMapper().writeValueAsString(generateDownloadUrlVO);
-        Mockito.doReturn(urls).when(mockStorageApiService)
-                .generateDataStorageItemUrl(ID, generateDownloadUrlVO.getPaths(),
-                                            generateDownloadUrlVO.getPermissions(), generateDownloadUrlVO.getHours());
-
-        final MvcResult mvcResult = performRequest(post(String.format(GENERATE_URL, ID)).content(content));
-
-        Mockito.verify(mockStorageApiService)
-                .generateDataStorageItemUrl(ID, generateDownloadUrlVO.getPaths(),
-                                            generateDownloadUrlVO.getPermissions(), generateDownloadUrlVO.getHours());
-        assertResponse(mvcResult, urls, DatastorageCreatorUtils.DOWNLOAD_FILE_URL_LIST_TYPE);
-    }
-
-    @Test
-    public void shouldFailGenerateDataStorageItemsUploadUrlsForUnauthorizedUser() {
-        performUnauthorizedRequest(post(String.format(GENERATE_UPLOAD_URL, ID)));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGenerateDataStorageItemsUploadUrls() throws Exception {
-        final DataStorageDownloadFileUrl url = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-        final List<DataStorageDownloadFileUrl> urls = Collections.singletonList(url);
-        final List<String> paths = Collections.singletonList(TEST);
-        final String content = getObjectMapper().writeValueAsString(paths);
-        Mockito.doReturn(urls).when(mockStorageApiService).generateDataStorageItemUploadUrl(ID, paths);
-
-        final MvcResult mvcResult = performRequest(post(String.format(GENERATE_UPLOAD_URL, ID)).content(content));
-
-        Mockito.verify(mockStorageApiService).generateDataStorageItemUploadUrl(ID, paths);
-        assertResponse(mvcResult, urls, DatastorageCreatorUtils.DOWNLOAD_FILE_URL_LIST_TYPE);
-    }
-
-    @Test
-    public void shouldFailRestoreFileVersionForUnauthorizedUser() {
-        performUnauthorizedRequest(post(String.format(RESTORE_VERSION_URL, ID)));
     }
 
     @Test
@@ -806,69 +398,6 @@ public class DataStorageControllerTest extends AbstractControllerTest {
 
         Mockito.verify(mockStorageApiService).deleteDataStorageObjectTags(ID, TEST, setTags, TEST);
         assertResponse(mvcResult, TAGS, STRING_STRING_MAP_TYPE);
-    }
-
-    @Test
-    public void shouldFailGetDataStorageItemsWithTagsForUnauthorizedUser() {
-        performUnauthorizedRequest(get(TAGS_LIST_URL, ID));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageFileWithTags() {
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(ID_PARAM, ID_AS_STRING);
-        params.add(PATH, TEST);
-        Mockito.doReturn(file).when(mockStorageApiService).getDataStorageItemWithTags(ID, TEST, false);
-
-        final MvcResult mvcResult = performRequest(get(String.format(TAGS_LIST_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemWithTags(ID, TEST, false);
-        assertResponse(mvcResult, file, DatastorageCreatorUtils.DATA_STORAGE_FILE_TYPE);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageFileOwnerWithTags() {
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(ID_PARAM, ID_AS_STRING);
-        params.add(PATH, TEST);
-        params.add(SHOW_VERSION, TRUE_AS_STRING);
-        Mockito.doReturn(file).when(mockStorageApiService).getDataStorageItemOwnerWithTags(ID, TEST, true);
-
-        final MvcResult mvcResult = performRequest(get(String.format(TAGS_LIST_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemOwnerWithTags(ID, TEST, true);
-        assertResponse(mvcResult, file, DatastorageCreatorUtils.DATA_STORAGE_FILE_TYPE);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageFolderWithTags() {
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(ID_PARAM, ID_AS_STRING);
-        params.add(PATH, TEST);
-        Mockito.doReturn(folder).when(mockStorageApiService).getDataStorageItemWithTags(ID, TEST, false);
-
-        final MvcResult mvcResult = performRequest(get(String.format(TAGS_LIST_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemWithTags(ID, TEST, false);
-        assertResponse(mvcResult, folder, DatastorageCreatorUtils.DATA_STORAGE_FOLDER_TYPE);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldGetDataStorageFolderOwnerWithTags() {
-        final MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add(ID_PARAM, ID_AS_STRING);
-        params.add(PATH, TEST);
-        params.add(SHOW_VERSION, TRUE_AS_STRING);
-        Mockito.doReturn(folder).when(mockStorageApiService).getDataStorageItemOwnerWithTags(ID, TEST, true);
-
-        final MvcResult mvcResult = performRequest(get(String.format(TAGS_LIST_URL, ID)).params(params));
-
-        Mockito.verify(mockStorageApiService).getDataStorageItemOwnerWithTags(ID, TEST, true);
-        assertResponse(mvcResult, folder, DatastorageCreatorUtils.DATA_STORAGE_FOLDER_TYPE);
     }
 
     @Test
